fix: Remove Connected and Unconnected profile
