+++ src/views/Nft/market/Profile/index.tsx
@@ -2,23 +2,48 @@ import React, { FC } from 'react'
 import { useWeb3React } from '@web3-react/core'
 import { useRouter } from 'next/router'
 import { isAddress } from 'utils'
-import { Flex, Text } from '@pancakeswap/uikit'
+import { Box, Flex, Text } from '@pancakeswap/uikit'
 import Page from 'components/Layout/Page'
 import { useTranslation } from 'contexts/Localization'
-import ConnectedProfile from './ConnectedProfile'
-import UnconnectedProfile from './UnconnectedProfile'
+import styled from 'styled-components'
+import useLastUpdated from 'hooks/useLastUpdated'
 import MarketPageHeader from '../components/MarketPageHeader'
 import ProfileHeader from './components/ProfileHeader'
 import NoNftsImage from '../components/Activity/NoNftsImage'
 import { NftMarketLayout } from '../Layout'
+import { useAchievementsForAddress, useProfileForAddress } from '../../../../state/profile/hooks'
+import useNftsForAddress from '../hooks/useNftsForAddress'
+import TabMenu from './components/TabMenu'
+
+const TabMenuWrapper = styled(Box)`
+  position: absolute;
+  bottom: 0;
+  left: 50%;
+  transform: translate(-50%, 0%);
+
+  ${({ theme }) => theme.mediaQueries.sm} {
+    left: auto;
+    transform: none;
+  }
+`
 
 const NftProfile: FC = ({ children }) => {
   const { account } = useWeb3React()
   const accountAddress = useRouter().query.accountAddress as string
   const { t } = useTranslation()
 
-  const isConnectedProfile = account?.toLowerCase() === accountAddress?.toLowerCase()
   const invalidAddress = !accountAddress || isAddress(accountAddress) === false
+
+  const { profile: profileHookState, isFetching: isProfileFetching } = useProfileForAddress(accountAddress)
+  const { profile } = profileHookState || {}
+  const { achievements, isFetching: isAchievementsFetching } = useAchievementsForAddress(accountAddress)
+  const { lastUpdated, setLastUpdated: refreshNfts } = useLastUpdated()
+  const { nfts: userNfts, isLoading: isNftLoading } = useNftsForAddress(
+    accountAddress,
+    profile,
+    isProfileFetching,
+    lastUpdated,
+  )
 
   if (invalidAddress) {
     return (
@@ -48,11 +73,22 @@ const NftProfile: FC = ({ children }) => {
 
   return (
     <>
-      {isConnectedProfile ? (
-        <ConnectedProfile>{children}</ConnectedProfile>
-      ) : (
-        <UnconnectedProfile>{children}</UnconnectedProfile>
-      )}
+      <MarketPageHeader position="relative">
+        <ProfileHeader
+          accountPath={account}
+          profile={profile}
+          achievements={achievements}
+          nftCollected={userNfts.length}
+          isProfileLoading={isProfileFetching}
+          isNftLoading={isNftLoading}
+          isAchievementsLoading={isAchievementsFetching}
+          onSuccess={refreshNfts}
+        />
+        <TabMenuWrapper>
+          <TabMenu />
+        </TabMenuWrapper>
+      </MarketPageHeader>
+      <Page style={{ minHeight: 'auto' }}>{children}</Page>
     </>
   )
 }
